[FEATURE] Notification email to parent thread posters etc.
Can also be used to send email to e.g. news article authors. The notification email is sent when the comment is published.

The TypoScript setting `plugin.tx_pwcomments.settings.sendMailOnCommentPublishTo` contains the recipients. By default it is configured to insert the register value `register : tx_pwcomments_threadEmails` (containing the authors of the parent comments) if the TypoScript constant `plugin.tx_pwcomments.settings.sendMailOnCommentPublishToThread` is true.